+++ test/vs-express/express/app.js
@@ -1,12 +1,14 @@
 var express = require("express"),
     path = require("path"),
+    url = require("url"),
     cookieParser = require("cookie-parser"),
     bodyParser = require("body-parser"),
 
     index = require("./routes/index"),
     users = require("./routes/users"),
 
-    app = express();
+    app = express(),
+    subdomain = "subdomain";
 
 app.set("strict routing", true);
 app.set("views", path.join(__dirname, "views"));
@@ -15,10 +17,10 @@ app.set("view engine", "pug");
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, "public")));
+app.use(url.format(path.join("/", subdomain, "/")), express.static(path.join(__dirname, "public")));
 
-app.use("/", index);
-app.use("/users", users);
+app.use(url.format(path.join("/", subdomain, "/")), index);
+app.use(url.format(path.join("/", subdomain, "/users")), users);
 
 app.use(function(req, res, next) {
     next(new Error());