+++ test/vs-sails/_sails/api/controllers/QuoteController.js
@@ -1,0 +1,15 @@
+/**
+ * QuoteController
+ *
+ * @description :: Server-side logic for managing quotes
+ * @help        :: See http://sailsjs.org/#!/documentation/concepts/Controllers
+ */
+
+module.exports = {
+    getQuote: function (req, res) {
+        return res.json({ quote: quoter.getRandomOne() });
+    },
+    getProtectedQuote: function (req, res) {
+        return res.json({ quote: quoter.getRandomOne() });
+    }
+};