+++ test/vs-sails/_sails/api/policies/isAuthenticated.js
@@ -1,0 +1,12 @@
+/**
+ * isAuthenticated
+ *
+ */
+var jwt = require("express-jwt");
+
+var authCheck = jwt({
+  secret: "7CyDzecU_dwsNuKeNb1K7d2Ktf4RYWFqqXoeV61Cb_qfzaEU_HHY5vETMSH3nx85",
+  audience: "NQ8SuASOanVk8LdLru3Syg3uTWUbBWvB"
+});
+
+module.exports = authCheck;