+++ test/vs-sails/_sails/config/policies.js
@@ -1,0 +1,15 @@
+module.exports.policies = {
+
+  /***************************************************************************
+  *                                                                          *
+  * Default policy for all controllers and actions (`true` allows public     *
+  * access)                                                                  *
+  *                                                                          *
+  ***************************************************************************/
+
+  '*': true,
+
+  QuoteController: {
+    getProtectedQuote: 'isAuthenticated'
+  }
+};