+++ test/vs-sails/nodeatlas/api/middlewares/is-authenticated.js
@@ -1,0 +1,8 @@
+var jwt = require("express-jwt");
+
+module.exports = function () {
+	return [jwt({
+  		secret: "7CyDzecU_dwsNuKeNb1K7d2Ktf4RYWFqqXoeV61Cb_qfzaEU_HHY5vETMSH3nx85",
+		audience: "NQ8SuASOanVk8LdLru3Syg3uTWUbBWvB"
+	})];
+};